\docType{package}
\name{praktikum-package}
\alias{praktikum}
\alias{praktikum-package}
\title{Lisamoodul 'praktikum'}
\description{
  \tabular{ll}{ Package: \tab praktikum\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2014-02-14\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }
}
\details{
  Kasulikud funktsioonid kvantitatiivsete mudelite kursuse
  (SHPH.00.004) jaoks / Functions used in the course
  "Quantitative methods in behavioural sciences"
  (SHPH.00.004), University of Tartu
}
\examples{
NULL
}
\author{
  Kenn Konstabel \email{kenn.konstabel AT tai.ee}
}
\keyword{package}

